requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydub
speechrecognition
gtts
tensorflow
pandas
numpy